Исправление ошибок. Сделал вызов создания большого фото из модуля gallery.js. Переименовал модуль full-size-photo.js на big-picture.js. Добавил атрибут data к к созданию фотоэлемента.
