+++ index.html
@@ -171,7 +171,7 @@
         </div>
 
         <!-- Кнопка для выхода из полноэкранного просмотра изображения -->
-        <button type="reset" class="big-picture__cancel  cancel" id="picture-cancel" tabindex="0">Закрыть</button>
+        <button type="reset" class="big-picture__cancel  cancel" id="picture-cancel">Закрыть</button>
       </div>
     </section>
   </main>
@@ -219,16 +219,16 @@
       </div>
     </section>
   </template>
-  
+
   <!--Комментарии к большому фото-->
   <template id="comments">
     <li class="social__comment">
       <img
           class="social__picture"
-          src
-          alt
-          width
-          height>
+          src=""
+          alt=""
+          width="35"
+          height="35">
       <p class="social__text">Комментарий</p>
     </li>
   </template>
